chore: use components export in blog example
