+++ examples/blog/site/pages/blog/hello-world.rocket.md
@@ -1,9 +1,9 @@
 ```js server
 /* START - Rocket auto generated - do not touch */
 export const sourceRelativeFilePath = 'blog/hello-world.rocket.md';
-import { html } from '../recursive.data.js';
+import { html, components } from '../recursive.data.js';
 import { layout } from './local.data.js';
-export { html, layout };
+export { html, components, layout };
 /* END - Rocket auto generated - do not touch */
 
 export const title = 'Hello world!';
@@ -12,12 +12,6 @@ export const author = 'Thomas Allmer (@daKmoR)';
 export const authorHref = 'https://twitter.com/daKmoR';
 export const value = 128;
 export const description = 'Just a Hello World Post!';
-
-// TODO: remove once lit-ssr can handle undefined values https://github.com/lit/lit/issues/2572
-export const heroImage = '';
-export const alt = '';
-
-import '../../src/components/blog-author.js';
 ```
 
 hey
